requirements.txt
streamlit
pandas
numpy
scikit-learn
scipy
opencv-python-headless
python-dotenv
matplotlib
seaborn